Rename `goal` submodule to `prove`
