+++ src/solve/mod.rs
@@ -9,7 +9,7 @@ pub mod match_program_clause;
 pub mod normalize;
 pub mod normalize_application;
 pub mod not_unify;
-pub mod goal;
+pub mod prove;
 pub mod solver;
 pub mod unify;
